requirements.txt
acres==0.5.0
alembic==1.17.0
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
marshmallow==4.0.1
moviepy==2.2.1
pillow==11.3.0
psycopg2==2.9.11
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
typing_extensions==4.15.0
Werkzeug==3.1.3
